requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.2
boto3==1.34.0
httpx==0.27.0
pytest==8.2.2
pytest-asyncio==0.23.7
mypy==1.10.0
types-requests==2.32.0.20240622
python-dotenv==1.0.1
mangum==0.17.0
requests==2.32.3
google-generativeai==0.7.2
pypdf==4.2.0
python-multipart==0.0.9
SQLAlchemy==2.0.31
psycopg[binary]==3.2.9
alembic==1.13.2

# Optional ML / LangChain extras moved to requirements-ml.txt to avoid numpy build issues on Python 3.13
